writing-practice/requirements.txt
streamlit
python-dotenv
watchdog
groq